+++ redditrepostsleuth/adminsvc/misc_admin_tasks.py
@@ -1,11 +1,16 @@
 from typing import NoReturn, List
 
 from praw import Reddit
+from sqlalchemy import func
 
+from redditrepostsleuth.core.config import Config
 from redditrepostsleuth.core.db.databasemodels import MonitoredSub
+from redditrepostsleuth.core.db.db_utils import get_db_engine
+from redditrepostsleuth.core.db.uow.sqlalchemyunitofworkmanager import SqlAlchemyUnitOfWorkManager
 from redditrepostsleuth.core.db.uow.unitofworkmanager import UnitOfWorkManager
 from redditrepostsleuth.core.logging import log
-from redditrepostsleuth.core.util.helpers import is_moderator, bot_has_permission
+from redditrepostsleuth.core.util.helpers import is_moderator, bot_has_permission, is_bot_banned
+from redditrepostsleuth.core.util.reddithelpers import get_reddit_instance
 
 
 def update_mod_status(uowm: UnitOfWorkManager, reddit: Reddit) -> NoReturn:
@@ -33,12 +38,27 @@ def update_mod_status(uowm: UnitOfWorkManager, reddit: Reddit) -> NoReturn:
             uow.commit()
 
 
+def update_ban_list(uowm: UnitOfWorkManager, reddit: Reddit) -> NoReturn:
+    """
+    Go through banned subs and see if we're still banned
+    :rtype: NoReturn
+    :param uowm: UnitOfWorkManager
+    :param reddit: Reddit
+    """
+    with uowm.start() as uow:
+        bans = uow.banned_subreddit.get_all()
+        for ban in bans:
+            subreddit = reddit.subreddit(ban.subreddit)
+            if is_bot_banned(subreddit):
+                log.info('Still banned on %s', ban.subreddit)
+                ban.last_checked = func.utc_timestamp()
+            else:
+                log.info('No longer banned on %s', ban.subreddit)
+                uow.banned_subreddit.remove(ban)
+            uow.commit()
+
 if __name__ == '__main__':
     config = Config(r'/home/barry/PycharmProjects/RedditRepostSleuth/sleuth_config.json')
     reddit = get_reddit_instance(config)
     uowm = SqlAlchemyUnitOfWorkManager(get_db_engine(config))
-    reddit_manager = RedditManager(reddit)
-    event_logger = EventLogging(config=config)
-    response_handler = ResponseHandler(reddit_manager, uowm, event_logger)
-    updater = SubredditConfigUpdater(uowm, reddit, response_handler, config)
-    updater.update_configs()
+    update_ban_list(uowm, reddit)